move idf to index time, and make everything lowercase
Moving the idf to index time, makes the search function a tiny bit
easier to read and manage.

transforming everything to lowercase is oversimplistic, but it is
required for ux in this app. We probably would want to make
transfomations specific to analyzers and also apply analyzers to search
queries if time permited.

these changes also include some linting